pyproject.toml
[tool.poetry]
name = "agent_asylum"
version = "0.1.0"
description = ""
authors = ["eightballer"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
open-aea-ledger-solana = "==1.64.0"
open-aea-ledger-cosmos = "==1.64.0"
open-aea-ledger-ethereum = "==1.64.0"
open-aea = "==1.64.0"
open-aea-cli-ipfs = "==1.64.0"
open-aea-test-autonomy = "==0.19.4"
open-autonomy = "==0.19.4"
autonomy-dev = {extras = ["all"], version = ">=0.2.64,<=0.2.138"}
python-telegram-bot = "<=21"

[tool.poetry.group.dev.dependencies]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]
